chore: update tests, documentation, and fix JSON marshalling/unmarshalling issues
- Improved test cases for enhanced coverage and reliability.
- Updated documentation to ensure clarity and uniformity with godoc standards.
- Minor fixes for marshalling and unmarshalling structs to JSON.
