+++ common/email_test.go
@@ -3,65 +3,11 @@ package common
 import (
 	"bytes"
 	"encoding/base64"
-	"encoding/json"
-	"strings"
 	"testing"
 
 	"github.com/darkrockmountain/gomail/sanitizer"
 	"github.com/stretchr/testify/assert"
 )
-
-func TestValidateEmail(t *testing.T) {
-	tests := []struct {
-		email    string
-		expected string
-	}{
-		{"test@example.com", "test@example.com"},
-		{"test@domain_name.com", "test@domain_name.com"},
-		{"test@domain-name.com", "test@domain-name.com"},
-		{"test@subdomain.example.com", "test@subdomain.example.com"},
-		{"test_name@subdomain.example.com", "test_name@subdomain.example.com"},
-		{"test.name@subdomain.example.com", "test.name@subdomain.example.com"},
-		{"test-name@subdomain.example.com", "test-name@subdomain.example.com"},
-		{"  test@example.com  ", "test@example.com"},
-		{"invalid-email", ""},
-		{"test@.com", ""},
-		{"@example.com", ""},
-		{"test@com", ""},
-		{"test@com.", ""},
-		{"test@sub.example.com", "test@sub.example.com"},
-		{"test+alias@example.com", "test+alias@example.com"},
-		{"test.email@example.com", "test.email@example.com"},
-		{"test-email@example.com", "test-email@example.com"},
-	}
-
-	for _, test := range tests {
-		t.Run(test.email, func(t *testing.T) {
-			result := ValidateEmail(test.email)
-			assert.Equal(t, test.expected, result)
-		})
-	}
-}
-
-func TestValidateEmailSlice(t *testing.T) {
-	tests := []struct {
-		emails   []string
-		expected []string
-	}{
-		{[]string{"test@example.com"}, []string{"test@example.com"}},
-		{[]string{"test@example.com", "invalid-email"}, []string{"test@example.com"}},
-		{[]string{" test@example.com ", "test2@example.com"}, []string{"test@example.com", "test2@example.com"}},
-		{[]string{"invalid-email", "@example.com"}, []string{}},
-		{[]string{"test@example.com", "test2@sub.example.com"}, []string{"test@example.com", "test2@sub.example.com"}},
-	}
-
-	for _, test := range tests {
-		t.Run(strings.Join(test.emails, ","), func(t *testing.T) {
-			result := ValidateEmailSlice(test.emails)
-			assert.Equal(t, test.expected, result)
-		})
-	}
-}
 
 func TestEmailMessageGetters(t *testing.T) {
 	message := EmailMessage{
@@ -353,176 +299,6 @@ func TestAddsEmailMessageToNils(t *testing.T) {
 		assert.Equal(t, html, email.html)
 		assert.Equal(t, []Attachment{attachment}, email.attachments)
 	})
-}
-
-func TestMarshalJSONCustom(t *testing.T) {
-	t.Run("Marshal EmailMessage with attachments", func(t *testing.T) {
-		email := &EmailMessage{
-			from:    "sender@example.com",
-			to:      []string{"recipient@example.com"},
-			cc:      []string{"cc@example.com"},
-			bcc:     []string{"bcc@example.com"},
-			replyTo: "replyto@example.com",
-			subject: "Subject",
-			text:    "This is the email content.",
-			html:    "<p>This is the email content.</p>",
-			attachments: []Attachment{
-				{filename: "attachment1.txt", content: []byte("content1")},
-			},
-			maxAttachmentSize: 1024,
-		}
-		jsonData, err := json.Marshal(email)
-		assert.Nil(t, err)
-		expected := `{"from":"sender@example.com","to":["recipient@example.com"],"cc":["cc@example.com"],"bcc":["bcc@example.com"],"replyTo":"replyto@example.com","subject":"Subject","text":"This is the email content.","html":"<p>This is the email content.</p>","attachments":[{"filename":"attachment1.txt","content":"Y29udGVudDE="}]}`
-		assert.JSONEq(t, expected, string(jsonData))
-	})
-
-	t.Run("Marshal EmailMessage without attachments", func(t *testing.T) {
-		email := &EmailMessage{
-			from:              "sender@example.com",
-			to:                []string{"recipient@example.com"},
-			cc:                []string{"cc@example.com"},
-			bcc:               []string{"bcc@example.com"},
-			replyTo:           "replyto@example.com",
-			subject:           "Subject",
-			text:              "This is the email content.",
-			html:              "<p>This is the email content.</p>",
-			maxAttachmentSize: 1024,
-		}
-		jsonData, err := json.Marshal(email)
-		assert.Nil(t, err)
-		expected := `{"from":"sender@example.com","to":["recipient@example.com"],"cc":["cc@example.com"],"bcc":["bcc@example.com"],"replyTo":"replyto@example.com","subject":"Subject","text":"This is the email content.","html":"<p>This is the email content.</p>"}`
-		assert.JSONEq(t, expected, string(jsonData))
-	})
-}
-
-func TestUnmarshalJSONCustom(t *testing.T) {
-	t.Run("Unmarshal EmailMessage with attachments", func(t *testing.T) {
-		jsonData := `{
-			"from": "sender@example.com",
-			"to": ["recipient@example.com"],
-			"cc": ["cc@example.com"],
-			"bcc": ["bcc@example.com"],
-			"replyTo": "replyto@example.com",
-			"subject": "Subject",
-			"text": "This is the email content.",
-			"html": "<p>This is the email content.</p>",
-			"attachments": [{"filename": "attachment1.txt", "content": "Y29udGVudDE="}]
-		}`
-		var email EmailMessage
-		err := json.Unmarshal([]byte(jsonData), &email)
-		assert.Nil(t, err)
-		assert.Equal(t, "sender@example.com", email.from)
-		assert.Equal(t, []string{"recipient@example.com"}, email.to)
-		assert.Equal(t, []string{"cc@example.com"}, email.cc)
-		assert.Equal(t, []string{"bcc@example.com"}, email.bcc)
-		assert.Equal(t, "replyto@example.com", email.replyTo)
-		assert.Equal(t, "Subject", email.subject)
-		assert.Equal(t, "This is the email content.", email.text)
-		assert.Equal(t, "<p>This is the email content.</p>", email.html)
-		expectedAttachment := Attachment{filename: "attachment1.txt", content: []byte("content1")}
-		assert.Equal(t, []Attachment{expectedAttachment}, email.attachments)
-	})
-
-	t.Run("Unmarshal EmailMessage without attachments", func(t *testing.T) {
-		jsonData := `{
-			"from": "sender@example.com",
-			"to": ["recipient@example.com"],
-			"cc": ["cc@example.com"],
-			"bcc": ["bcc@example.com"],
-			"replyTo": "replyto@example.com",
-			"subject": "Subject",
-			"text": "This is the email content.",
-			"html": "<p>This is the email content.</p>"
-		}`
-		var email EmailMessage
-		err := json.Unmarshal([]byte(jsonData), &email)
-		assert.Nil(t, err)
-		assert.Equal(t, "sender@example.com", email.from)
-		assert.Equal(t, []string{"recipient@example.com"}, email.to)
-		assert.Equal(t, []string{"cc@example.com"}, email.cc)
-		assert.Equal(t, []string{"bcc@example.com"}, email.bcc)
-		assert.Equal(t, "replyto@example.com", email.replyTo)
-		assert.Equal(t, "Subject", email.subject)
-		assert.Equal(t, "This is the email content.", email.text)
-		assert.Equal(t, "<p>This is the email content.</p>", email.html)
-		assert.Nil(t, email.attachments)
-	})
-}
-
-func TestMarshalJSONEdgeCases(t *testing.T) {
-	t.Run("nil EmailMessage", func(t *testing.T) {
-		var email *EmailMessage
-		result, err := json.Marshal(email)
-		assert.Nil(t, err)
-		assert.Equal(t, "null", string(result))
-	})
-
-	t.Run("nil Attachment", func(t *testing.T) {
-		var attachment *Attachment
-		result, err := json.Marshal(attachment)
-		assert.Nil(t, err)
-		assert.Equal(t, "null", string(result))
-	})
-}
-
-func TestUnmarshalJSONEdgeCases(t *testing.T) {
-	t.Run("empty JSON EmailMessage", func(t *testing.T) {
-		jsonData := `{}`
-		var email EmailMessage
-		err := json.Unmarshal([]byte(jsonData), &email)
-		assert.Nil(t, err)
-		assert.Equal(t, "", email.from)
-		assert.Equal(t, "", email.subject)
-		assert.Equal(t, "", email.text)
-		assert.Equal(t, "", email.html)
-	})
-
-	t.Run("invalid JSON EmailMessage", func(t *testing.T) {
-		// Example of JSON data that will cause the unmarshaling to fail
-		// 'to' should be an array, not a string
-		invalidJSONData := `{
-        "from": "sender@example.com",
-        "to": "invalid_recipient@example.com",
-        "cc": ["cc@example.com"],
-        "bcc": ["bcc@example.com"],
-        "replyTo": "replyto@example.com",
-        "subject": "Subject",
-        "text": "This is the email content.",
-        "html": "<p>This is the email content.</p>",
-        "attachments": [{"filename": "attachment1.txt", "content": "Y29udGVudDE="}]
-    }`
-
-		var email EmailMessage
-		err := json.Unmarshal([]byte(invalidJSONData), &email)
-		assert.Error(t, err)
-	})
-
-	t.Run("invalid JSON Attachment", func(t *testing.T) {
-		// Example of JSON data that will cause the unmarshaling to fail
-		// 'filename' should be an string, not an integer
-		jsonData := `{"filename": 123456789, "content": "invalid_base64"}`
-		var attachment Attachment
-		err := json.Unmarshal([]byte(jsonData), &attachment)
-		assert.Error(t, err)
-	})
-
-	t.Run("empty JSON Attachment", func(t *testing.T) {
-		jsonData := `{}`
-		var attachment Attachment
-		err := json.Unmarshal([]byte(jsonData), &attachment)
-		assert.Nil(t, err)
-		assert.Equal(t, "", attachment.filename)
-		assert.Equal(t, []byte{}, attachment.content)
-	})
-
-	t.Run("invalid base64 content Attachment", func(t *testing.T) {
-		jsonData := `{"filename": "file.txt", "content": "invalid_base64"}`
-		var attachment Attachment
-		err := json.Unmarshal([]byte(jsonData), &attachment)
-		assert.NotNil(t, err)
-	})
-
 }
 
 func TestIsHTMLEdgeCases(t *testing.T) {
